+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- npm: pnpm-workspace.yaml support for catalog dependencies (pnpm 9+)
+
 ### Changed
 
 ### Removed